% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBikeData.R
\name{checkBikeData}
\alias{checkBikeData}
\title{Check correct format of bike sharing data}
\usage{
checkBikeData(data)
}
\arguments{
\item{data}{pre-processed bike sharing data frame
with columns Hour_chunks, Is_weekend,
Is_holiday, Season, Min_temp, Max_temp,
Min_humidity, Max_humidity, Wind_speed,
Rain_or_snow, Date, Bike_count}
}
\description{
This is a helper function to ensure that the bike
sharing dataset used as input to a function is
properly formatted. Throws warnings if formatting
is incorrect.
}
