% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkModelGLMM.R
\name{checkModelGLMM}
\alias{checkModelGLMM}
\title{Check correct format of negative binomial mixed model object}
\usage{
checkModelGLMM(glmm.model)
}
\arguments{
\item{glmm.model}{trained negative binomial mixed model
(output of MCEM_algorithm() function)}
}
\description{
This is a helper function to ensure that the negative binomial mixed model
object used as input to a function is properly formatted. Throws warnings
if formatting is incorrect.
}
