% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkModelRF.R
\name{checkModelRF}
\alias{checkModelRF}
\title{Check correct format of random forest model object}
\usage{
checkModelRF(rf.model)
}
\arguments{
\item{rf.model}{trained random forest caret object
(output of train_random_forest() function)}
}
\description{
This is a helper function to ensure that the random forest trained caret
object used as input to a function is properly formatted. Throws warnings
if formatting is incorrect.
}
