% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkParamsMCEM.R
\name{checkParamsMCEM}
\alias{checkParamsMCEM}
\title{Check correct format of MCEM parameters}
\usage{
checkParamsMCEM(
  beta_initial,
  theta_initial,
  s2gamma_initial,
  M,
  burn.in,
  tol,
  maxit,
  trace
)
}
\arguments{
\item{beta_initial}{initial beta values (vector of 14 numeric values)}

\item{theta_initial}{initial theta value (single numeric value)}

\item{s2gamma_initial}{initial value for s2gamma (single numeric value)}

\item{M}{length of chain (single numeric count value)}

\item{burn.in}{the number of beginning iteration samples to discard
(single numeric count value)}

\item{tol}{tolerance for convergence (single numeric value)}

\item{maxit}{maximum number of iterations to run (single numeric count value)}

\item{trace}{level of detailed output (single numeric value)}
}
\description{
This is a helper function to ensure that the correct parameters (other
than data) are given to the MCEM algorithm function before running.
}
