% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeBinomialGLMM.R
\name{g.sim}
\alias{g.sim}
\title{Add prior u to random walk}
\usage{
g.sim(u)
}
\arguments{
\item{u}{prior}
}
\value{
u + sampled random walk value
}
\description{
This helper function simulates a sample from the proposal
density g, which is the prior plus the value from a random walk
}
