% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeBinomialGLMM.R
\name{MCEM_algorithm}
\alias{MCEM_algorithm}
\title{Run MCEM algorithm to estimate NB GLMM bike count parameters}
\usage{
MCEM_algorithm(
  data,
  beta_initial = c(8.3, 1.5, 1.5, -0.25, -0.5, 0, 0, -0.25, 0, 0, 0, 0, 0, -0.25),
  theta_initial = 10,
  s2gamma_initial = 0.2,
  M = 1000,
  burn.in = 200,
  tol = 10^-4,
  maxit = 100,
  trace = 0
)
}
\arguments{
\item{data}{pre-processed bike sharing data frame
with columns Bike_count, Hour_chunks,
Max_temp, Rain_or_snow, and Date}

\item{beta_initial}{initial beta values (vector of 14 numeric values)}

\item{theta_initial}{initial theta value (single numeric value)}

\item{s2gamma_initial}{initial value for s2gamma (single numeric value)}

\item{M}{length of chain (single numeric count value)}

\item{burn.in}{the number of beginning iteration samples to discard
(single numeric count value)}

\item{tol}{tolerance for convergence (single numeric value)}

\item{maxit}{maximum number of iterations to run (single numeric count value)}

\item{trace}{level of detailed output (single numeric value)}
}
\value{
a list containing beta, theta, s2gamma, qfunction, and date random
effect estimates, as well as epsilon and the iteration count
}
\description{
This function runs the Monte Carlo EM algorithm as a means to estimate the
parameters of a negative binomial general mixed model used to predict bike
counts from date, time, and weather data. Hourly chunks, maximum daily
temperature, and presence of rain or snow are modeled as fixed effects, and
date is modeled as a random effect. The Metropolis Hasting algorithm is used
to draw samples of the random effect and the Nelder-Mead algorithm is used
to calculate the M-step.
}
