% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{plot_rf_importance}
\alias{plot_rf_importance}
\title{Plot random forest variable importance}
\usage{
plot_rf_importance(data, mtry = 11)
}
\arguments{
\item{data}{pre-processed bike sharing data frame
with columns Bike_count, Hour_chunks,
Max_temp, Rain_or_snow, and Date}

\item{mtry}{the mtry value from the model$bestTune
parameter of the trained random forest
model using the trainRandomForest(data)}
}
\value{
a ggplot object
}
\description{
This function plots the amount of increase in node
purity vs. the amount of increase in MSE that a
each variable in a random forest model provides
}
