% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{plotRandomForestImportance}
\alias{plotRandomForestImportance}
\title{Plot random forest variable importance}
\usage{
plotRandomForestImportance(model)
}
\arguments{
\item{model}{an object of class randomForest}
}
\value{
a ggplot object
}
\description{
This function plots the amount of increase in node
purity vs. the amount of increase in MSE that a
each variable in a random forest model provides
}
