% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeBinomialGLMM.R
\name{Qi}
\alias{Qi}
\title{Helper function to calculate the Q-function for the ith subject}
\usage{
Qi(yi, xi, thetat, betat, s2gammat, gammai, burn_in)
}
\description{
Helper function to calculate the Q-function for the ith subject
}
