% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeBinomialGLMM.R
\name{r.walk}
\alias{r.walk}
\title{Perform random walk}
\usage{
r.walk()
}
\value{
sampled random value
}
\description{
This helper function performs a random walk, drawing a
sample from a uniform distribution between -1/4 and 1/4
}
