% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAssessment.R
\name{rfModelFit}
\alias{rfModelFit}
\title{Assess random forest model fit}
\usage{
rfModelFit(model, data, scale_to_seoul_mean = "no")
}
\arguments{
\item{data}{pre-processed bike sharing data frame
with columns Bike_count, Hour_chunks,
Max_temp, Rain_or_snow, and Date}

\item{scale_to_seoul_mean}{"yes" or "no". If "yes",
scales the bike count mean of the input
data to the Seoul data bike count mean}
}
\value{
an object of class train containing the
fit random forest model
}
\description{
This function assesses the fit of a trained random forest
model on a given dataset
}
