% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeBinomialGLMM.R
\name{sample.gamma.posterior.all}
\alias{sample.gamma.posterior.all}
\title{Helper function to draw M samples from the posterior for gammai}
\usage{
sample.gamma.posterior.all(data, M, maxit, thetat, betat, s2gammat, trace = 0)
}
\description{
Helper function to draw M samples from the posterior for gammai
}
