% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeBinomialGLMM.R
\name{sample.gamma.posterior.i}
\alias{sample.gamma.posterior.i}
\title{Helper function to run random walk sampler for subject i}
\usage{
sample.gamma.posterior.i(yi, xi, M, maxit, thetat, betat, s2gammat, trace = 0)
}
\description{
Helper function to run random walk sampler for subject i
}
