% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{train_random_forest}
\alias{train_random_forest}
\title{Train a random forest model on bike sharing data}
\usage{
train_random_forest(data)
}
\arguments{
\item{data}{pre-processed bike sharing data frame
with columns Hour_chunks, Is_weekend,
Is_holiday, Season, Min_temp, Max_temp,
Min_humidity, Max_humidity, Wind_speed,
Rain_or_snow, Date, Bike_count}
}
\value{
an object of class train containing the
fit random forest model
}
\description{
This function trains a random forest model on bike
sharing data using the Hour_chunks, Max_temp, and
Rain_or_snow predictors.
}
