% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{trainRandomForest}
\alias{trainRandomForest}
\title{Train a random forest model on bike sharing data}
\usage{
trainRandomForest(data)
}
\arguments{
\item{data}{pre-processed bike sharing data frame
with columns Bike_count, Hour_chunks,
Max_temp, Rain_or_snow, and Date}
}
\value{
an object of class train containing the
fit random forest model
}
\description{
This function trains a random forest model on bike
sharing data using the Hour_chunks, Max_temp, and
Rain_or_snow predictors.
}
