% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBikeData.R
\name{checkBikeData}
\alias{checkBikeData}
\title{Check correct format of bike sharing data}
\usage{
checkBikeData(data)
}
\arguments{
\item{data}{pre-processed bike sharing data frame
with columns Bike_count, Hour_chunks,
Max_temp, Rain_or_snow, and Date}
}
\description{
This is a helper function to ensure that the bike
sharing dataset used as input to a function is
properly formatted. Throws warnings if formatting
is incorrect.
}
