% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeBinomialGLMM.R
\name{MCEM_algorithm}
\alias{MCEM_algorithm}
\title{Run MCEM algorithm to estimate NB GLMM bike count parameters}
\usage{
MCEM_algorithm(
  beta_initial = c(7.49, 0.8, 1.06, -0.15, -0.23, -0.3, -0.18, -0.72, -0.01, 0.05, 0,
    0, 0.01, -0.48),
  theta_initial = 4.542,
  s2gamma_initial = 0.2,
  M = 1000,
  burn.in = 200,
  tol = 10^-5,
  maxit = 100,
  data = seoul,
  trace = 0
)
}
\arguments{
\item{beta_initial}{initial beta values}

\item{s2gamma_initial}{initial value for s2gamma}

\item{M}{length of chain}

\item{burn.in}{the number of beginning iteration samples to discard}

\item{tol}{tolerance for convergence}

\item{maxit}{maximum number of iterations to run}

\item{data}{pre-processed bike sharing data frame
with columns Bike_count, Hour_chunks,
Max_temp, Rain_or_snow, and Date}

\item{trace}{level of detailed output}

\item{theta_inital}{initial theta value}
}
\value{
a list containing beta, theta, s2gamma, qfunction, and date random
effect estimates, as well as epsilon and the iteration count
}
\description{
This function runs the Monte Carlo EM algorithm as a means to estimate the
parameters of a negative binomial general mixed model used to predict bike
counts from date, time, and weather data. Hourly chunks, maximum daily
temperature, and presence of rain or snow are modeled as fixed effects, and
date is modeled as a random effect. The Metropolis Hasting algorithm is used
to draw samples of the random effect and the Nelder-Mead algorithm is used
to calculate the M-step.
}
