% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeBinomialGLMM.R
\name{Q}
\alias{Q}
\title{Helper function for calculating the Q-function}
\usage{
Q(data, thetat, betat, s2gammat, samples, burn_in, logs2gammat = F)
}
\description{
Helper function for calculating the Q-function
}
