% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfmodelAssessment.R
\name{rf_model_fit}
\alias{rf_model_fit}
\title{Assess random forest model fit}
\usage{
rf_model_fit(model, data, scale_seoul_mean = "no")
}
\arguments{
\item{model}{trained random forest model for assessment}

\item{data}{pre-processed bike sharing data frame
with columns Hour_chunks, Is_weekend, Is_holiday,
Season, Min_temp, Max_temp, Min_humidity, Max_humidity,
Wind_speed, Rain_or_snow, Date, Bike_count}

\item{scale_seoul_mean}{"yes" or "no". If "yes",
scales the bike count mean of the input
data to the Seoul data bike count mean}
}
\value{
a data frame summarizing the RMSE, MAE, and R2 value of the fitted
rf model to the data
}
\description{
This function assesses the fit of a trained random forest
model on a given dataset
}
